refactor: 채널 리스트
